+++ android/src/main/java/com/reactnativestripesdk/addresssheet/AddressSheetViewManager.kt
@@ -1,4 +1,4 @@
-package com.reactnativestripesdk
+package com.reactnativestripesdk.addresssheet
 
 import com.facebook.react.bridge.ReadableArray
 import com.facebook.react.bridge.ReadableMap
@@ -6,8 +6,6 @@ import com.facebook.react.common.MapBuilder
 import com.facebook.react.uimanager.SimpleViewManager
 import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.annotations.ReactProp
-import com.stripe.android.paymentsheet.addresselement.AddressDetails
-import com.stripe.android.paymentsheet.addresselement.AddressLauncher
 
 class AddressSheetViewManager : SimpleViewManager<AddressSheetView>() {
   override fun getName() = "AddressSheetView"
